requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
python-multipart
email-validator
torch
numpy
opencv-python-headless
Pillow
rosbags
google-cloud-aiplatform
google-cloud-firestore
python-dotenv
